backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
pyyaml==6.0.1
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
jsonschema==4.21.1
jmespath==1.0.1
openai==1.3.0 